move `display_and_edit_input_data()` to layout elements module
